Clean shutdown. Export embedded broker connection client id.
* Seperate EmbeddedBrokerService and EmbeddedBrokerServiceImpl
* Make persistence file name building more readable
